animal_adoption_platform/animal_posts_service/requirements.txt
grpcio
grpcio-tools
Flask
requests
SQLAlchemy
